Move the `platform` packages into separate directory
