--- arc/internal/header/platform/platform.go
@@ -1,6 +1,0 @@
-package platform
-
-// Возвращает размеры терминала
-func GetTerminalSize() (int, int, error) {
-	return getTerminalSize()
-}
