+++ arc/internal/platform/compress_read_darwin.go
