+++ arc/internal/platform/compress_read_windows.go
