+++ arc/internal/platform/header_windows.go
