+++ go.mod
@@ -3,4 +3,4 @@ module github.com/gh0st17/archiver
 go 1.23.6
 
 require golang.org/x/term v0.29.0
-require golang.org/x/sys v0.30.0 // indirect
+require golang.org/x/sys v0.30.0
